Support for multiple managed iframes on the same page
Follow-up for 2220857926f5e4068d9d6deaf7c29d6cdacb5023.

This adds support for multiple iframes on the same page, where some (or all) of them are managed through the GTM script.

This works by looking though the iframes present on the parent page to find an iframe corresponding to the message sender window (a.k.a. `event.source` field), then working with the selected iframe.
Since the frames request initialization and other wise initiate communication themselves, the mechanism does not require any extra configuration on the parent page.
Among other things, this means that the `gtm-resize-enable` flag is no longer required. Support for it has been dropped.